move `transform()` to separate file
