+++ reporting/transformer_example/src/transform.py
@@ -1,0 +1,15 @@
+# Some kind of transform -------------------------------------------------------
+def transform(record):
+    '''
+    Parameters
+    ----------
+    record : dict
+        the raw data from VHS in a malleable dict format
+        
+    Returns
+    -------
+    transformed_record : dict
+        record with necessary transformations applied, ready to be formatted for
+        elasticsearch ingestion
+    '''
+    return record
